+++ TerrariumGardenTech.Repositories/Entity/BlogCategory.cs
@@ -7,7 +7,7 @@ namespace TerrariumGardenTech.Repositories.Entity;
 
 public partial class BlogCategory
 {
-    public long CategoryId { get; set; }
+    public int CategoryId { get; set; }
 
     public string CategoryName { get; set; }
